+++ data/src/main/java/edu/nextstep/camp/calculator/data/CalculateRepository.kt
@@ -1,6 +1,6 @@
-package edu.nextstep.camp.calculator
+package edu.nextstep.camp.calculator.data
 
-import edu.nextstep.camp.calculator.local.History
+import edu.nextstep.camp.calculator.data.local.History
 
 interface CalculateRepository {
     val historyAll: List<History>
