+++ data/src/main/java/edu/nextstep/camp/calculator/data/LocalCalculateRepository.kt
@@ -1,7 +1,7 @@
-package edu.nextstep.camp.calculator
+package edu.nextstep.camp.calculator.data
 
-import edu.nextstep.camp.calculator.local.History
-import edu.nextstep.camp.calculator.local.HistoryDao
+import edu.nextstep.camp.calculator.data.local.History
+import edu.nextstep.camp.calculator.data.local.HistoryDao
 
 internal class LocalCalculateRepository(private val historyDao: HistoryDao) : CalculateRepository {
     override val historyAll: List<History>
